+++ module/CLI/src/Command/Visit/GetNonOrphanVisitsCommand.php
@@ -1,0 +1,46 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shlinkio\Shlink\CLI\Command\Visit;
+
+use Shlinkio\Shlink\Common\Paginator\Paginator;
+use Shlinkio\Shlink\Common\Util\DateRange;
+use Shlinkio\Shlink\Core\Entity\Visit;
+use Shlinkio\Shlink\Core\Model\VisitsParams;
+use Shlinkio\Shlink\Core\ShortUrl\Helper\ShortUrlStringifierInterface;
+use Shlinkio\Shlink\Core\Visit\VisitsStatsHelperInterface;
+use Symfony\Component\Console\Input\InputInterface;
+
+class GetNonOrphanVisitsCommand extends AbstractVisitsListCommand
+{
+    public const NAME = 'visit:non-orphan';
+
+    public function __construct(
+        VisitsStatsHelperInterface $visitsHelper,
+        private readonly ShortUrlStringifierInterface $shortUrlStringifier,
+    ) {
+        parent::__construct($visitsHelper);
+    }
+
+    protected function doConfigure(): void
+    {
+        $this
+            ->setName(self::NAME)
+            ->setDescription('Returns the list of non-orphan visits.');
+    }
+
+    protected function getVisitsPaginator(InputInterface $input, DateRange $dateRange): Paginator
+    {
+        return $this->visitsHelper->nonOrphanVisits(new VisitsParams($dateRange));
+    }
+
+    /**
+     * @return array<string, string>
+     */
+    protected function mapExtraFields(Visit $visit): array
+    {
+        $shortUrl = $visit->getShortUrl();
+        return $shortUrl === null ? [] : ['shortUrl' => $this->shortUrlStringifier->stringify($shortUrl)];
+    }
+}